refactor: SearchSuggestions·Search 중복 제거 → Articles::Search로 통합
- Articles::SearchSuggestions.suggest + similarity_order → Articles::Search 로 이동
- Search::VectorMath/ReciprocalRankFusion/MaximalMarginalRelevance 중복 제거
- app/functions/search/ 디렉토리 제거
- 테스트 통합: search_test.rb (20 runs, 0 failures)
